Check functions passed to exported functions

Internal helper function that check whether a function passed as an argument meets the requirements of package methods.

test_fn_req_args() checks whether the function has only the expected number of required arguments, i.e., arguments without default values. Defaults to checking for a single required argument.

test_fn_num_out() checks whether a function accepting a numeric vector input returns a numeric vector output of the same length as the input, with finite non-missing values \(\geq\) 0.0.

Usage

test_func_req_args(func, n_req_args = 1)

test_func_num_out(func, n = 10)

Arguments

func

A function. This is expected to be a function evaluating the density of a distribution at numeric values.

n_req_args

The number of required arguments, i.e., arguments without default values.

n

The number of elements over which to evaluate the function func. Defaults to 10, and func is evaluated over seq(n).

Value

A logical for whether the function func fulfils conditions specified in the respective checks.